+++ pennylane/templates/embeddings.py
@@ -18,14 +18,70 @@ used at the beginning of a circuit.
 """
 #pylint: disable-msg=too-many-branches,too-many-arguments,protected-access
 import numpy as np
-
-from pennylane.ops import RX, RY, RZ, BasisState, Squeezing, Displacement, QubitStateVector
+from pennylane.ops import RX, RY, RZ, CNOT, Hadamard, BasisState, Squeezing, Displacement, QubitStateVector
 from pennylane.templates.utils import (_check_shape, _check_no_variable, _check_wires,
-                                       _check_hyperp_is_in_options, _check_type)
+                                       _check_hyperp_is_in_options, _check_type,
+                                       _check_number_of_layers, _get_shape)
 from pennylane.variable import Variable
 
 
 TOLERANCE = 1e-3
+
+
+def _qaoa_feature_encoding_hamiltonian(features, n_features, wires):
+    """Implements the encoding Hamiltonian of the QAOA embedding.
+
+    Args:
+        features (array): array of features to encode
+        n_features (int): number of features to encode
+    """
+    for idx, w in enumerate(wires):
+        # Either feed in feature
+        if idx < n_features:
+            RX(features[idx], wires=w)
+        # or a Hadamard
+        else:
+            Hadamard(wires=w)
+
+
+def _qaoa_ising_hamiltonian(weights, wires, local_fields, l):
+    """Implements the Ising-like Hamiltonian of the QAOA embedding.
+
+    Args:
+        weights (array): array of weights
+        wires (Sequence[int] or int): `n` qubit indices that the template acts on
+        local_fields (str): gate implementing the local field
+        l (int): layer index
+    """
+    # trainable "Ising" ansatz
+    if len(wires) == 1:
+        local_fields(weights[l, 0], wires=wires[0])
+
+    elif len(wires) == 2:
+        # ZZ coupling
+        CNOT(wires=[wires[0], wires[1]])
+        RZ(2 * weights[l, 0], wires=wires[0])
+        CNOT(wires=[wires[0], wires[1]])
+
+        # local fields
+        for i, _ in enumerate(wires):
+            local_fields(weights[l, i + 1], wires=wires[i])
+
+    else:
+        for i, _ in enumerate(wires):
+            if i < len(wires) - 1:
+                # ZZ coupling
+                CNOT(wires=[wires[i], wires[i + 1]])
+                RZ(2 * weights[l, i], wires=wires[i])
+                CNOT(wires=[wires[i], wires[i + 1]])
+            else:
+                # ZZ coupling to enforce periodic boundary condition
+                CNOT(wires=[wires[i], wires[0]])
+                RZ(2 * weights[l, i], wires=wires[i])
+                CNOT(wires=[wires[i], wires[0]])
+        # local fields
+        for i, _ in enumerate(wires):
+            local_fields(weights[l, len(wires) + i], wires=wires[i])
 
 
 def AmplitudeEmbedding(features, wires, pad=None, normalize=False):
@@ -295,6 +351,190 @@ def BasisEmbedding(features, wires):
     BasisState(features, wires=wires)
 
 
+def QAOAEmbedding(features, weights, wires, local_field='Y'):
+    r"""
+    Encodes :math:`N` features into :math:`n>N` qubits, using a layered, trainable quantum
+    circuit that is inspired by the QAOA ansatz.
+
+    A single layer applies two circuits or "Hamiltonians": The first encodes the features, and the second is
+    a variational ansatz inspired by a 1-dimensional Ising model. The feature-encoding circuit associates features with
+    the angles of :class:`RX` rotations. The Ising ansatz consists of trainable two-qubit ZZ interactions
+    :math:`e^{-i \alpha \sigma_z \otimes \sigma_z}`,
+    and trainable local fields :math:`e^{-i \frac{\beta}{2} \sigma_{\mu}}`, where :math:`\sigma_{\mu}`
+    can be chosen to be :math:`\sigma_{x}`, :math:`\sigma_{y}` or :math:`\sigma_{z}`
+    (default choice is :math:`\sigma_{y}` or the ``RY`` gate), and :math:`\alpha, \beta` are adjustable gate parameters.
+
+    The number of features has to be smaller or equal to the number of qubits. If there are fewer features than
+    qubits, the feature-encoding rotation is replaced by a Hadamard gate.
+
+    The argument ``weights`` contains an array of the :math:`\alpha, \beta` parameters for each layer.
+    The number of layers :math:`L` is derived from the first dimension of ``weights``, which has the following
+    shape:
+
+    * :math:`(L, )`, if the embedding acts on a single wire,
+    * :math:`(L, 3)`, if the embedding acts on two wires,
+    * :math:`(L, 2n)` else.
+
+    After the :math:`L` th layer, another set of feature-encoding :class:`RX` gates is applied.
+
+    This is an example for the full embedding circuit using 2 layers, 3 features, 4 wires, and ``RY`` local fields:
+
+    |
+
+    .. figure:: ../../_static/qaoa_layers.png
+        :align: center
+        :width: 60%
+        :target: javascript:void(0);
+
+    |
+
+    .. note::
+        ``QAOAEmbedding`` supports gradient computations with respect to both the ``features`` and the ``weights``
+        arguments. Note that trainable parameters need to be passed to the quantum node as positional arguments.
+
+    Args:
+        features (array): array of features to encode
+        weights (array): array of weights
+        wires (Sequence[int] or int): `n` qubit indices that the template acts on
+        local_field (str): type of local field used, one of ``'X'``, ``'Y'``, or ``'Z'``
+
+    Raises:
+        ValueError: if inputs do not have the correct format
+
+    .. UsageDetails::
+
+        The QAOA embedding encodes an :math:`n`-dimensional feature vector into at most :math:`n` qubits. The
+        embedding applies layers of a circuit, and each layer is defined by a set of weight parameters.
+
+        .. code-block:: python
+
+            import pennylane as qml
+            from pennylane.templates import QAOAEmbedding
+
+            dev = qml.device('default.qubit', wires=2)
+
+            @qml.qnode(dev)
+            def circuit(weights, f=None):
+                QAOAEmbedding(features=f, weights=weights, wires=range(2))
+                return qml.expval(qml.PauliZ(0))
+
+            features = [1., 2.]
+            layer1 = [0.1, -0.3, 1.5]
+            layer2 = [3.1, 0.2, -2.8]
+            weights = [layer1, layer2]
+
+            print(circuit(weights, f=features))
+
+        **Using parameter initialization functions**
+
+        The initial weight parameters can alternatively be generated by utility functions from the
+        ``pennylane.init`` module, for example using the function :func:`~.qaoa_embedding_normal`:
+
+        .. code-block:: python
+
+            from pennylane.init import qaoa_embedding_normal
+            weights = qaoa_embedding_normal(n_layers=2, n_wires=2, mean=0, std=0.2)
+
+
+        **Training the embedding**
+
+        The embedding is typically trained with respect to a given cost. For example, one can train it to
+        minimize the PauliZ expectation of the first qubit:
+
+        .. code-block:: python
+
+            o = GradientDescentOptimizer()
+            for i in range(10):
+                weights = o.step(lambda w : circuit(w, f=features), weights)
+                print("Step ", i, " weights = ", weights)
+
+
+        **Training the features**
+
+        In principle, also the features are trainable, which means that gradients with respect to feature values
+        can be computed. To train both weights and features, they need to be passed to the qnode as
+        positional arguments. If the built-in optimizer is used, they have to be merged to one input:
+
+        .. code-block:: python
+
+            @qml.qnode(dev)
+            def circuit2(pars):
+                weights = pars[0]
+                features = pars[1]
+                QAOAEmbedding(features=features, weights=weights, wires=range(2))
+                return qml.expval(qml.PauliZ(0))
+
+
+            features = [1., 2.]
+            weights = [[0.1, -0.3, 1.5], [3.1, 0.2, -2.8]]
+            pars = [weights, features]
+
+            o = GradientDescentOptimizer()
+            for i in range(10):
+                pars = o.step(circuit2, pars)
+                print("Step ", i, " weights = ", pars[0], " features = ", pars[1])
+
+        **Local Fields**
+
+        While by default, ``RY`` gates are used as local fields, one may also choose ``local_field='Z'`` or
+        ``local_field='X'`` as hyperparameters of the embedding.
+
+        .. code-block:: python
+
+            @qml.qnode(dev)
+            def circuit(weights, f=None):
+                QAOAEmbedding(features=f, weights=weights, wires=range(2), local_field='Z')
+                return qml.expval(qml.PauliZ(0))
+
+        Choosing ``'Z'`` fields implements a QAOAEmbedding where the second Hamiltonian is a
+        1-dimensional Ising model.
+
+    """
+    #############
+    # Input checks
+    wires, n_wires = _check_wires(wires)
+
+    n_features = _get_shape(features)[0]
+    msg = "QAOAEmbedding cannot process more features than number of qubits {};" \
+          "got {}.".format(n_wires, len(features))
+    _check_shape(features, (n_wires,), bound='max', msg=msg)
+
+    msg = "Option for local field not known. Has to be one of ``'X'``, ``'Y'``, or ``'Z'``."
+    _check_hyperp_is_in_options(local_field, ['X', 'Y', 'Z'], msg=msg)
+    if local_field == 'Z':
+        local_fields = RZ
+    elif local_field == 'X':
+        local_fields = RX
+    else:
+        local_fields = RY
+
+    repeat = _check_number_of_layers([weights])
+
+    weights = np.array(weights)
+    weights_shape = weights.shape
+    if n_wires == 1:
+        msg = "QAOAEmbedding with 1 qubit and {} layers requires weight " \
+              "array of shape {}; got {}".format(repeat, (repeat, 1), weights_shape)
+        _check_shape(weights, (repeat, 1), msg=msg)
+    elif n_wires == 2:
+        msg = "QAOAEmbedding with 2 qubits and {} layers requires weight " \
+              "array of shape {}; got {}".format(repeat, (repeat, 3), weights_shape)
+        _check_shape(weights, (repeat, 3), msg=msg)
+    else:
+        msg = "QAOAEmbedding with {} qubits and {} layers requires weight " \
+              "array of shape {}; got {}".format(n_wires, repeat, (repeat, 2*n_wires), weights_shape)
+        _check_shape(weights, (repeat, 2*n_wires), msg=msg)
+    #####################
+
+    for l in range(repeat):
+        # apply alternating Hamiltonians
+        _qaoa_feature_encoding_hamiltonian(features, n_features, wires)
+        _qaoa_ising_hamiltonian(weights, wires, local_fields, l)
+
+    # repeat the feature encoding once more at the end
+    _qaoa_feature_encoding_hamiltonian(features, n_features, wires)
+
+
 def DisplacementEmbedding(features, wires, method='amplitude', c=0.1):
     r"""Encodes :math:`N` features into the displacement amplitudes :math:`r` or phases :math:`\phi` of :math:`M` modes,
      where :math:`N\leq M`.
@@ -392,6 +632,6 @@ def SqueezingEmbedding(features, wires, method='amplitude', c=0.1):
 
 
 embeddings = {"AngleEmbedding", "AmplitudeEmbedding", "BasisEmbedding", "DisplacementEmbedding",
-              "SqueezingEmbedding"}
+              "QAOAEmbedding", "SqueezingEmbedding"}
 
 __all__ = list(embeddings)